requirements.txt
Django
Django-timescaledb
dj-database-url
jupyter
requests
psycopg[binary]
python-decouple
pytz
celery
Redis
Django-celery-beat
Django-celery-results
decouple
Django-admin-interface
Django-admin-rangefilter
openai
Djangorestframework
